Cleaned up SocketSet, made it use unsigned int as index/size type.
Minor 64bit compatiblity fixes.


git-svn-id: svn://rakshasa.no/libtorrent/trunk/libtorrent@444 e378c898-3ddf-0310-93e7-cc216c733640

+++ src/data/memory_chunk.cc
@@ -53,7 +53,7 @@ MemoryChunk::MemoryChunk(char* ptr, char* begin, char* end, int prot, int flags)
   if (page_align() >= m_pagesize)
     throw internal_error("MemoryChunk::MemoryChunk(...) received an page alignment >= page size");
 
-  if ((uint32_t)ptr % m_pagesize)
+  if ((ptrdiff_t)ptr % m_pagesize)
     throw internal_error("MemoryChunk::MemoryChunk(...) is not aligned to a page");
 }
